LayoutTests/fast/dom/Node/insertBefore-no-op-mutationobserver-expected.txt

Tests that calls of insertBefore() lead to mutation events even if they are no-ops.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Mutation observer was notified
PASS wasMutationEventListenerCalled is true
PASS Mutation observer was notified
PASS wasMutationEventListenerCalled is true
PASS successfullyParsed is true

TEST COMPLETE
